+++ src/compas_timber/connections/__init__.py
@@ -18,10 +18,11 @@ from .t_lap import TLapJoint
 from .x_lap import XLapJoint
 from .x_notch import XNotchJoint
 from .t_dovetail import TDovetailJoint
-from .t_tenon_mortise import TenonMortiseJoint
+from .t_tenon_mortise import TTenonMortiseJoint
+from .l_tenon_mortise import LTenonMortiseJoint
 from .ball_node import BallNodeJoint
 from .y_butt import YButtJoint
-from .oligina import OliGinaJoint
+from .oligina import TOliGinaJoint
 from .utilities import beam_ref_side_incidence
 from .utilities import beam_ref_side_incidence_with_vector
 from .utilities import point_centerline_towards_joint
@@ -64,9 +65,10 @@ __all__ = [
     "find_neighboring_elements",
     "TDovetailJoint",
     "BallNodeJoint",
-    "TenonMortiseJoint",
+    "TTenonMortiseJoint",
+    "LTenonMortiseJoint",
     "YButtJoint",
-    "OliGinaJoint",
+    "TOliGinaJoint",
     "beam_ref_side_incidence",
     "beam_ref_side_incidence_with_vector",
     "point_centerline_towards_joint",